admin_service/requirements/requirements.txt
# FastAPI and web server
fastapi==0.110.2
uvicorn==0.29.0
starlette==0.37.2

# Database
sqlalchemy==2.0.41
asyncpg==0.29.0
alembic==1.13.1

# Templates and static files
jinja2==3.1.3
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
pydantic==2.11.7
pydantic_core==2.33.2
